В миграцию migration_2_3 добавлена проверка на существование поля pc_id в таблице table_users.

+++ app/src/main/java/com/mu/pclist/data/DatabaseMigration.kt
@@ -23,6 +23,11 @@ val migration_1_2: Migration = object : Migration(1, 2) {
 
 val migration_2_3: Migration = object : Migration(2, 3) {
     override fun migrate(db: SupportSQLiteDatabase) {
-        db.execSQL("ALTER TABLE table_users DROP COLUMN pc_id")
+        val cursor = db.query(
+            "SELECT * FROM table_users"
+        )
+        if (cursor.columnNames.contains("pc_id")) {
+            db.execSQL("ALTER TABLE table_users DROP COLUMN pc_id")
+        }
     }
 }
